+++ src/main/java/com/taxicalls/trip/resources/exceptions/DriverNotFoundException.java
@@ -8,7 +8,7 @@ public class DriverNotFoundException extends RuntimeException {
 
     private static final long serialVersionUID = 1L;
 
-    public DriverNotFoundException(Integer id) {
+    public DriverNotFoundException(Long id) {
         super("No such entity: " + id);
     }
 }
